# apps/api/src/security/enums.ts
export enum AuditEvent {
  USER_REGISTERED = 'USER_REGISTERED',
  AUTHENTICATION_SUCCEEDED = 'AUTHENTICATION_SUCCEEDED',
  AUTHENTICATION_FAILED = 'AUTHENTICATION_FAILED',
  TOKEN_REFRESHED = 'TOKEN_REFRESHED',
  SESSION_LOGGED_OUT = 'SESSION_LOGGED_OUT',
  CONSENT_GRANTED = 'CONSENT_GRANTED',
  CONSENT_REVOKED = 'CONSENT_REVOKED',
  IDENTITY_VIEWED = 'IDENTITY_VIEWED',
  RISK_EVALUATION_REQUESTED = 'RISK_EVALUATION_REQUESTED',
  REPORT_REQUESTED = 'REPORT_REQUESTED'
}

export enum Role {
  CONSUMER = 'CONSUMER',
  SME_OWNER = 'SME_OWNER',
  BUSINESS_USER = 'BUSINESS_USER',
  COOPERATIVE_ADMIN = 'COOPERATIVE_ADMIN',
  ADMIN = 'ADMIN',
  SUPER_ADMIN = 'SUPER_ADMIN'
}

export enum ConsentStatus {
  PENDING = 'PENDING', GRANTED = 'GRANTED', REVOKED = 'REVOKED', EXPIRED = 'EXPIRED'
}

export enum RiskRecommendation {
  REVIEW_RECOMMENDED = 'REVIEW_RECOMMENDED',
  ADDITIONAL_INFORMATION_RECOMMENDED = 'ADDITIONAL_INFORMATION_RECOMMENDED',
  MANUAL_ASSESSMENT_RECOMMENDED = 'MANUAL_ASSESSMENT_RECOMMENDED'
}