# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API process is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check API process health",
        "tags": [
          "system"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Register a consumer account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Authenticate without disclosing account existence",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh token and issue a new access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/me": {
      "get": {
        "operationId": "me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return the authenticated user and RBAC role",
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "CrediPass Nigeria API",
    "description": "Consent-led credit passport infrastructure. Sprint 0 contract only.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "HealthDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ],
            "example": "ok"
          },
          "service": {
            "type": "string",
            "example": "credipass-api"
          },
          "version": {
            "type": "string",
            "example": "0.1.0"
          }
        },
        "required": [
          "status",
          "service",
          "version"
        ]
      },
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string",
            "example": "demo-request-001"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-26T00:00:00.000Z"
          }
        },
        "required": [
          "correlationId",
          "timestamp"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/HealthDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "new.user@example.test"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "example": "Correct-Horse-47!"
          },
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "New User"
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "CONSUMER",
          "SME_OWNER",
          "BUSINESS_USER",
          "COOPERATIVE_ADMIN",
          "ADMIN",
          "SUPER_ADMIN"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user_demo_001"
          },
          "displayName": {
            "type": "string",
            "example": "New User"
          },
          "maskedEmail": {
            "type": "string",
            "example": "ne***@example.test"
          },
          "role": {
            "example": "CONSUMER",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        },
        "required": [
          "id",
          "displayName",
          "maskedEmail",
          "role"
        ]
      },
      "SessionSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "session_demo_001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "expiresAt"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT access token."
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque rotating refresh token. Store securely."
          },
          "expiresIn": {
            "type": "number",
            "example": 900
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "user",
          "session"
        ]
      },
      "SuccessMetaDto": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string",
            "example": "request-001"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "correlationId",
          "timestamp"
        ]
      },
      "AuthSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/AuthResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMetaDto"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "AuthErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "message": {
            "type": "string",
            "example": "Invalid email or password."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorMetaDto": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string",
            "example": "request-001"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string",
            "example": "/auth/login"
          }
        },
        "required": [
          "correlationId",
          "timestamp",
          "path"
        ]
      },
      "AuthErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/AuthErrorDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ErrorMetaDto"
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "new.user@example.test"
          },
          "password": {
            "type": "string",
            "example": "Correct-Horse-47!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "session-id.secret"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "loggedOut": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "loggedOut"
        ]
      },
      "LogoutSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/LogoutResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMetaDto"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "CurrentUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          }
        },
        "required": [
          "user",
          "session"
        ]
      },
      "CurrentUserSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/CurrentUserResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMetaDto"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      }
    }
  }
}